Remove unused lib/ from .gitignore and rename utils/ back to lib/
The lib/ gitignore entry was for TypeScript build output that is no longer
used since the viewer imports language source .ts files directly via Vite.
