+++ .gitignore
@@ -148,6 +148,3 @@ dist
 #!! ERROR: typescript is undefined. Use list command to see defined gitignore types !!#
 
 # End of https://www.toptal.com/developers/gitignore/api/typescript
-
-# TypeScript build output
-lib/
